+++ Firestore/Source/API/FIRDocumentSnapshot.m
@@ -26,6 +26,7 @@
 #import "Firestore/Source/Model/FSTDocumentKey.h"
 #import "Firestore/Source/Model/FSTFieldValue.h"
 #import "Firestore/Source/Model/FSTPath.h"
+#import "Firestore/Source/Util/FSTAssert.h"
 #import "Firestore/Source/Util/FSTUsageValidation.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -100,23 +101,15 @@ NS_ASSUME_NONNULL_BEGIN
   return _cachedMetadata;
 }
 
-- (NSDictionary<NSString *, id> *)data {
+- (nullable NSDictionary<NSString *, id> *)data {
   return [self dataWithOptions:[FIRSnapshotOptions defaultOptions]];
 }
 
-- (NSDictionary<NSString *, id> *)dataWithOptions:(FIRSnapshotOptions *)options {
-  FSTDocument *document = self.internalDocument;
-
-  if (!document) {
-    FSTThrowInvalidUsage(
-        @"NonExistentDocumentException",
-        @"Document '%@' doesn't exist. "
-        @"Check document.exists to make sure the document exists before calling document.data.",
-        self.internalKey);
-  }
-
-  return [self convertedObject:[self.internalDocument data]
-                       options:[FSTFieldValueOptions optionsForSnapshotOptions:options]];
+- (nullable NSDictionary<NSString *, id> *)dataWithOptions:(FIRSnapshotOptions *)options {
+  return self.internalDocument == nil
+             ? nil
+             : [self convertedObject:[self.internalDocument data]
+                             options:[FSTFieldValueOptions optionsForSnapshotOptions:options]];
 }
 
 - (nullable id)valueForField:(id)field {
@@ -135,8 +128,10 @@ NS_ASSUME_NONNULL_BEGIN
   }
 
   FSTFieldValue *fieldValue = [[self.internalDocument data] valueForPath:fieldPath.internalValue];
-  return [self convertedValue:fieldValue
-                      options:[FSTFieldValueOptions optionsForSnapshotOptions:options]];
+  return fieldValue == nil
+             ? nil
+             : [self convertedValue:fieldValue
+                            options:[FSTFieldValueOptions optionsForSnapshotOptions:options]];
 }
 
 - (nullable id)objectForKeyedSubscript:(id)key {
@@ -186,6 +181,36 @@ NS_ASSUME_NONNULL_BEGIN
     [result addObject:[self convertedValue:value options:options]];
   }];
   return result;
+}
+
+@end
+
+@interface FIRQueryDocumentSnapshot ()
+
+- (instancetype)initWithFirestore:(FIRFirestore *)firestore
+                      documentKey:(FSTDocumentKey *)documentKey
+                         document:(FSTDocument *)document
+                        fromCache:(BOOL)fromCache NS_DESIGNATED_INITIALIZER;
+
+@end
+
+@implementation FIRQueryDocumentSnapshot
+
+- (instancetype)initWithFirestore:(FIRFirestore *)firestore
+                      documentKey:(FSTDocumentKey *)documentKey
+                         document:(FSTDocument *)document
+                        fromCache:(BOOL)fromCache {
+  self = [super initWithFirestore:firestore
+                      documentKey:documentKey
+                         document:document
+                        fromCache:fromCache];
+  return self;
+}
+
+- (NSDictionary<NSString *, id> *)data {
+  NSDictionary<NSString *, id> *data = [super data];
+  FSTAssert(data, @"Document in a QueryDocumentSnapshot should exist");
+  return data;
 }
 
 @end